pyproject.toml
[project]
name = "megano-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "user",email = "ilya.melnichuk.2007@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=6.0.4,<7.0.0)",
    "psycopg (>=3.3.3,<4.0.0)",
    "djangorestframework (>=3.17.1,<4.0.0)",
    "gunicorn (>=25.3.0,<26.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "setuptools (>=82.0.1,<83.0.0)",
    "sentry-sdk (==2.35.0)",
    "pillow (>=12.2.0,<13.0.0)",
    "django-prometheus (>=2.4.1,<3.0.0)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "pytest (>=9.0.3,<10.0.0)",
    "pytest-django (>=4.12.0,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
